fixing build on windows
Ignore-this: d4d3aaf49fe210c101264e72652dfd48

darcs-hash:b85fce507e2a52119b1badc6dd3399326d2289de

+++ src/gui/app/chat.cpp
@@ -56,8 +56,8 @@ namespace fire
 
             namespace
             {
-                const size_t MIN_WIDTH = 400;
-                const size_t MIN_HEIGHT = 200;
+                const int MIN_WIDTH = 400;
+                const int MIN_HEIGHT = 200;
                 const std::string MESSAGE = "m";
                 const std::string JOINED = "j";
             }
